requirements.txt
aiohttp==3.8.1
marshmallow-dataclass==8.5.3
pyyaml==5.4.1
